apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic-settings==2.8.1
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
faiss-cpu==1.10.0
langgraph==1.0.4
langchain-core==1.1.2
PyYAML==6.0.3
